Extract weight section of log app into its own component
The goal here is just to make it easier to read the code / keep the
`log.vue` component from blowing up into a huge, complicated mess.
